packages/embeddings/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "embeddings"
version = "0.1.0"
description = "Embedding generation and vector store for OT threat intelligence"
requires-python = ">=3.12,<3.13"
dependencies = [
    "core",
    "ingestion",
    "boto3>=1.34",
    "numpy>=1.26",
    "faiss-cpu>=1.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/embeddings"]
